# kratos/api/system/v1/menu_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v3.19.4
// source: system/v1/menu.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationMenuServiceCreateMenu = "/api.system.v1.MenuService/CreateMenu"
const OperationMenuServiceDeleteMenu = "/api.system.v1.MenuService/DeleteMenu"
const OperationMenuServiceGetMenuById = "/api.system.v1.MenuService/GetMenuById"
const OperationMenuServiceGetMenuList = "/api.system.v1.MenuService/GetMenuList"
const OperationMenuServiceGetMenuTree = "/api.system.v1.MenuService/GetMenuTree"
const OperationMenuServiceGetUserMenuTree = "/api.system.v1.MenuService/GetUserMenuTree"
const OperationMenuServiceUpdateMenu = "/api.system.v1.MenuService/UpdateMenu"

type MenuServiceHTTPServer interface {
	CreateMenu(context.Context, *MenuItem) (*MessageReply, error)
	DeleteMenu(context.Context, *MenuIdRequest) (*MessageReply, error)
	GetMenuById(context.Context, *MenuIdRequest) (*MenuItem, error)
	GetMenuList(context.Context, *GetMenuListRequest) (*MenuListReply, error)
	GetMenuTree(context.Context, *GetMenuTreeRequest) (*MenuTreeReply, error)
	GetUserMenuTree(context.Context, *GetUserMenuTreeRequest) (*MenuTreeReply, error)
	UpdateMenu(context.Context, *UpdateMenuRequest) (*MessageReply, error)
}

func RegisterMenuServiceHTTPServer(s *http.Server, srv MenuServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/menu/user/tree", _MenuService_GetUserMenuTree0_HTTP_Handler(srv))
	r.GET("/api/v1/menu/tree", _MenuService_GetMenuTree0_HTTP_Handler(srv))
	r.GET("/api/v1/menu", _MenuService_GetMenuList0_HTTP_Handler(srv))
	r.GET("/api/v1/menu/{id}", _MenuService_GetMenuById0_HTTP_Handler(srv))
	r.POST("/api/v1/menu", _MenuService_CreateMenu0_HTTP_Handler(srv))
	r.PUT("/api/v1/menu/{id}", _MenuService_UpdateMenu0_HTTP_Handler(srv))
	r.DELETE("/api/v1/menu/{id}", _MenuService_DeleteMenu0_HTTP_Handler(srv))
}

func _MenuService_GetUserMenuTree0_HTTP_Handler(srv MenuServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUserMenuTreeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuServiceGetUserMenuTree)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserMenuTree(ctx, req.(*GetUserMenuTreeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MenuTreeReply)
		return ctx.Result(200, reply)
	}
}

func _MenuService_GetMenuTree0_HTTP_Handler(srv MenuServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMenuTreeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuServiceGetMenuTree)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMenuTree(ctx, req.(*GetMenuTreeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MenuTreeReply)
		return ctx.Result(200, reply)
	}
}

func _MenuService_GetMenuList0_HTTP_Handler(srv MenuServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMenuListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuServiceGetMenuList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMenuList(ctx, req.(*GetMenuListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MenuListReply)
		return ctx.Result(200, reply)
	}
}

func _MenuService_GetMenuById0_HTTP_Handler(srv MenuServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in MenuIdRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuServiceGetMenuById)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMenuById(ctx, req.(*MenuIdRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MenuItem)
		return ctx.Result(200, reply)
	}
}

func _MenuService_CreateMenu0_HTTP_Handler(srv MenuServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in MenuItem
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuServiceCreateMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateMenu(ctx, req.(*MenuItem))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MessageReply)
		return ctx.Result(200, reply)
	}
}

func _MenuService_UpdateMenu0_HTTP_Handler(srv MenuServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMenuRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuServiceUpdateMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMenu(ctx, req.(*UpdateMenuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MessageReply)
		return ctx.Result(200, reply)
	}
}

func _MenuService_DeleteMenu0_HTTP_Handler(srv MenuServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in MenuIdRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMenuServiceDeleteMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteMenu(ctx, req.(*MenuIdRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MessageReply)
		return ctx.Result(200, reply)
	}
}

type MenuServiceHTTPClient interface {
	CreateMenu(ctx context.Context, req *MenuItem, opts ...http.CallOption) (rsp *MessageReply, err error)
	DeleteMenu(ctx context.Context, req *MenuIdRequest, opts ...http.CallOption) (rsp *MessageReply, err error)
	GetMenuById(ctx context.Context, req *MenuIdRequest, opts ...http.CallOption) (rsp *MenuItem, err error)
	GetMenuList(ctx context.Context, req *GetMenuListRequest, opts ...http.CallOption) (rsp *MenuListReply, err error)
	GetMenuTree(ctx context.Context, req *GetMenuTreeRequest, opts ...http.CallOption) (rsp *MenuTreeReply, err error)
	GetUserMenuTree(ctx context.Context, req *GetUserMenuTreeRequest, opts ...http.CallOption) (rsp *MenuTreeReply, err error)
	UpdateMenu(ctx context.Context, req *UpdateMenuRequest, opts ...http.CallOption) (rsp *MessageReply, err error)
}

type MenuServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewMenuServiceHTTPClient(client *http.Client) MenuServiceHTTPClient {
	return &MenuServiceHTTPClientImpl{client}
}

func (c *MenuServiceHTTPClientImpl) CreateMenu(ctx context.Context, in *MenuItem, opts ...http.CallOption) (*MessageReply, error) {
	var out MessageReply
	pattern := "/api/v1/menu"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMenuServiceCreateMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MenuServiceHTTPClientImpl) DeleteMenu(ctx context.Context, in *MenuIdRequest, opts ...http.CallOption) (*MessageReply, error) {
	var out MessageReply
	pattern := "/api/v1/menu/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMenuServiceDeleteMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MenuServiceHTTPClientImpl) GetMenuById(ctx context.Context, in *MenuIdRequest, opts ...http.CallOption) (*MenuItem, error) {
	var out MenuItem
	pattern := "/api/v1/menu/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMenuServiceGetMenuById))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MenuServiceHTTPClientImpl) GetMenuList(ctx context.Context, in *GetMenuListRequest, opts ...http.CallOption) (*MenuListReply, error) {
	var out MenuListReply
	pattern := "/api/v1/menu"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMenuServiceGetMenuList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MenuServiceHTTPClientImpl) GetMenuTree(ctx context.Context, in *GetMenuTreeRequest, opts ...http.CallOption) (*MenuTreeReply, error) {
	var out MenuTreeReply
	pattern := "/api/v1/menu/tree"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMenuServiceGetMenuTree))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MenuServiceHTTPClientImpl) GetUserMenuTree(ctx context.Context, in *GetUserMenuTreeRequest, opts ...http.CallOption) (*MenuTreeReply, error) {
	var out MenuTreeReply
	pattern := "/api/v1/menu/user/tree"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMenuServiceGetUserMenuTree))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MenuServiceHTTPClientImpl) UpdateMenu(ctx context.Context, in *UpdateMenuRequest, opts ...http.CallOption) (*MessageReply, error) {
	var out MessageReply
	pattern := "/api/v1/menu/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMenuServiceUpdateMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}