# native/docs/swagger/swagger.yaml
basePath: /
definitions:
  github_com_force-c_nai-tizi_internal_domain_model.Menu:
    properties:
      component:
        description: 组件路径
        type: string
      createBy:
        description: 创建人
        type: integer
      createdTime:
        type: string
      icon:
        description: 菜单图标
        type: string
      id:
        description: 菜单ID（使用分布式ID）
        type: integer
      isCache:
        description: 是否缓存：0否 1是
        type: integer
      isFrame:
        description: 是否外链：0否 1是
        type: integer
      menuName:
        description: 菜单名称
        type: string
      menuType:
        description: 菜单类型：0目录 1菜单 2按钮
        type: integer
      parentId:
        description: 父菜单ID（0表示根菜单）
        type: integer
      path:
        description: 路由地址
        type: string
      perms:
        description: '权限标识（例如: user.create, user.*, *）'
        type: string
      query:
        description: 路由参数
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 显示顺序
        type: integer
      status:
        description: 状态：0正常 1停用
        type: integer
      updateBy:
        description: 更新人
        type: integer
      updatedTime:
        type: string
      visible:
        description: 显示状态：0显示 1隐藏
        type: integer
    type: object
...
  github_com_force-c_nai-tizi_internal_domain_response.Response:
    description: API 统一响应格式
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 响应消息
        example: success
        type: string
    type: object
...
  github_com_force-c_nai-tizi_internal_service.MenuTree:
    properties:
      children:
        items:
          $ref: '#/definitions/github_com_force-c_nai-tizi_internal_service.MenuTree'
        type: array
      component:
        description: 组件路径
        type: string
      createBy:
        description: 创建人
        type: integer
      createdTime:
        type: string
      icon:
        description: 菜单图标
        type: string
      id:
        description: 菜单ID（使用分布式ID）
        type: integer
      isCache:
        description: 是否缓存：0否 1是
        type: integer
      isFrame:
        description: 是否外链：0否 1是
        type: integer
      menuName:
        description: 菜单名称
        type: string
      menuType:
        description: 菜单类型：0目录 1菜单 2按钮
        type: integer
      parentId:
        description: 父菜单ID（0表示根菜单）
        type: integer
      path:
        description: 路由地址
        type: string
      perms:
        description: '权限标识（例如: user.create, user.*, *）'
        type: string
      query:
        description: 路由参数
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 显示顺序
        type: integer
      status:
        description: 状态：0正常 1停用
        type: integer
      updateBy:
        description: 更新人
        type: integer
      updatedTime:
        type: string
      visible:
        description: 显示状态：0显示 1隐藏
        type: integer
    type: object
...
host: localhost:9009
info:
  contact:
    email: support@example.com
    name: 技术支持
    url: https://example.com/support
  description: Nai-Tizi RESTful API 接口文档，支持双 Token 认证机制
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: NTZ API 文档
  version: "1.0"
paths:
...
  /api/v1/menu:
    get:
      consumes:
      - application/json
      description: 获取所有菜单的列表形式
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_model.Menu'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
      security:
      - Bearer: []
      summary: 获取菜单列表
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 创建新的菜单项
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
      security:
      - Bearer: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /api/v1/menu/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
      security:
      - Bearer: []
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      consumes:
      - application/json
      description: 根据菜单ID获取菜单详细信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_model.Menu'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "404":
          description: 菜单不存在
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
      security:
      - Bearer: []
      summary: 获取菜单详情
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      description: 更新菜单信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
      security:
      - Bearer: []
      summary: 更新菜单
      tags:
      - 菜单管理
  /api/v1/menu/tree:
    get:
      consumes:
      - application/json
      description: 获取所有菜单的树形结构，用于菜单管理页面
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_force-c_nai-tizi_internal_service.MenuTree'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
      security:
      - Bearer: []
      summary: 获取菜单树
      tags:
      - 菜单管理
  /api/v1/menu/user/tree:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的菜单树，用于前端生成动态路由
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_force-c_nai-tizi_internal_service.MenuTree'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
      security:
      - Bearer: []
      summary: 获取用户菜单树
      tags:
      - 菜单管理
...
securityDefinitions:
  Bearer:
    description: '格式: "Bearer {access_token}"，AccessToken 从登录接口获取'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...